+++ src/org/cicirello/search/operators/permutations/WindowLimitedUndoableScrambleMutation.java
@@ -56,8 +56,7 @@ public final class WindowLimitedUndoableScrambleMutation extends UndoableScrambl
 	 * a default window limit of Integer.MAX_VALUE.
 	 */
 	public WindowLimitedUndoableScrambleMutation() {
-		super();
-		limit = Integer.MAX_VALUE;
+		this(Integer.MAX_VALUE);
 	}
 	
 	/**